+++ src/commands/index.ts
@@ -15,4 +15,5 @@ program
   .command("object:import", "Import an object from Ninox", {
     executableFile: "object-import",
   })
+  .command("deploy", "Deploy a Ninox SDK project", {executableFile: "deploy"})
   .parse(process.argv);
