--- src/credentials.ts
@@ -1,5 +1,0 @@
-export interface Credentials {
-  workspaceId: string;
-  domain: string;
-  apiKey: string;
-}
