+++ src/util/yaml.util.ts
@@ -5,3 +5,7 @@ export const createYamlDocument = (data: any) => {
   doc.contents = data;
   return doc;
 };
+
+export const parseYamlDocument = (yaml: string) => {
+  return YAML.parse(yaml);
+};
